requirements.txt
syncedlyrics
colorama
yt-dlp
python-vlc